feat: add typescript for board function
