+++ src/app.ts
@@ -2,7 +2,7 @@ import { FastifyInstance, RegisterOptions } from 'fastify';
 
 import { userRouter } from './resources/users/user.router';
 import { boardRouter } from './resources/boards/board.router';
-import { tasksRouter } from './resources/tasks/tasks.router';
+// import { tasksRouter } from './resources/tasks/tasks.router';
 
 export type Done = () => void;
 
@@ -13,7 +13,7 @@ export function app(
 ) {
   userRouter(fastify, options, done);
   boardRouter(fastify, options, done);
-  tasksRouter(fastify, options, done);
+  // tasksRouter(fastify, options, done);
 }
 
 // module.exports = app;
