--- src/resources/boards/board.memory.repository.js
@@ -1,3 +1,0 @@
-const items = [];
-
-module.exports = items;
