+++ src/resources/boards/board.memory.repository.ts
@@ -1,0 +1,4 @@
+// const items = [];
+
+// module.exports = items;
+export const items: Array<{ id: string; title: string; columns: [] }> = [];
